requirements.txt
python-telegram-bot==21.3
httpx>=0.28.1
python-dotenv==1.0.1
apscheduler==3.10.4
